+++ platform-queue/src/main/java/com/flow/platform/queue/PriorityQueueItem.java
@@ -21,7 +21,7 @@ package com.flow.platform.queue;
  */
 public interface PriorityQueueItem extends Comparable<PriorityQueueItem> {
 
-    Integer getPriority();
+    long getPriority();
 
     byte[] getBody();
 }
